# inventory/openapi.yaml
openapi: 3.1.0
info:
  title: DyingStar Inventory API
  version: 0.1.0
  description: >
    Ownership registry of DyingStar: who owns which goods. Goods are either **fungible**
    (a stack: type + quantity, e.g. 500 t of ore) or **unique** (an instance with a stable
    UUID, e.g. a specific truck). This service stores ownership and reservations only — it
    never tracks physical location and never moves goods by itself: the game server (or the
    market/mission service) confirms that an exchange really happened, then ownership moves.

    Three kinds of callers:

    * **Players** send a Keycloak JWT (`Authorization: Bearer`). With `AUTH_DEV_BYPASS=true`
      (never in production) the `X-Player-Id` / `X-Player-Name` headers are accepted instead.
    * **The game server, the market and the mission service** authenticate with a Keycloak
      service-account token (`client_credentials`) whose `azp` is an allowed client and whose
      `aud` targets `inventory-api`. Each `/api/internal/*` route requires a capability role
      (e.g. `inventory:hold`). Dev only (`INTERNAL_DEV_BYPASS=true`): the legacy
      `X-Internal-Key` header is accepted.

    Beyond goods, the service stores **POIs** (points of interest): a GPS position
    (`x`/`y`/`z`), optionally a zone (`radiusM`), scoped to a `system`/`scene`, owned by a
    player, NPC, corporation, political entity or the reserved `system`. POIs are shared
...
    `POST /api/internal/pois/resolve` to match its availability zones.

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/corporations/{corporationId}/inventory:
    get:
      tags: [corporations]
      summary: Corporation inventory (member only, membership checked in Social)
      parameters:
        - $ref: '#/components/parameters/corporationId'
      responses:
        '200':
          description: Inventory of the corporation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HolderInventory' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: Not a member of this corporation (`NOT_CORPORATION_MEMBER`), or Social not configured
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/corporations/{corporationId}/inventory/stacks/{goodType}:
    get:
      tags: [corporations]
      summary: One corporation stack (member only)
      parameters:
        - $ref: '#/components/parameters/corporationId'
        - $ref: '#/components/parameters/goodType'
      responses:
        '200':
          description: The stack
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StackView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ── POIs (points of interest / zones) ────────────────────────────────────────
...
  /api/corporations/{corporationId}/pois:
    get:
      tags: [corporations]
      summary: Corporation POIs (member only, membership checked in Social)
      parameters:
        - $ref: '#/components/parameters/corporationId'
      responses:
        '200':
          description: POIs owned by or granted to the corporation
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Poi' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/SocialUnavailable' }
  /api/corporations/{corporationId}/pois/{poiId}:
    get:
      tags: [corporations]
      summary: One POI of the corporation's scope (member only)
      parameters:
        - $ref: '#/components/parameters/corporationId'
        - $ref: '#/components/parameters/poiId'
      responses:
        '200':
          description: The POI and its grants
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PoiView' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/SocialUnavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak access token (RS256, validated against the realm JWKS). The player id is the
        `sub` claim. Dev only (`AUTH_DEV_BYPASS=true`): `X-Player-Id`, `X-Player-Name`,
        `X-Player-Roles` headers may replace the token.
    serviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Keycloak service-account access token (client_credentials, RS256). `azp` must be an
        allowed client (`INTERNAL_SERVICE_CLIENTS`) and `aud` must contain `inventory-api`;
        each route then needs its capability role. Dev only (`INTERNAL_DEV_BYPASS=true`): the
        legacy `X-Internal-Key` header is accepted instead.
  parameters:
...
    corporationId:
      name: corporationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    goodType:
      name: goodType
      in: path
      required: true
      schema: { type: string, maxLength: 128 }
...
    poiId:
      name: poiId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid player token / service token (`UNAUTHORIZED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Not a corporation member (`NOT_CORPORATION_MEMBER`) or missing role (`FORBIDDEN`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found (`NOT_FOUND`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    SocialUnavailable:
      description: >
        The Social service is unreachable or not configured while checking organization
        membership (`SOCIAL_NOT_CONFIGURED`, `SOCIAL_LOOKUP_FAILED`, `SOCIAL_AUTH_FAILED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error, message, status]
      properties:
        error:
          type: string
          description: Machine-readable code
          examples: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, SERVICE_FORBIDDEN, NOT_CORPORATION_MEMBER, NOT_FOUND, CONFLICT, INSUFFICIENT_GOODS, NOT_POI_OWNER, POI_ALREADY_SHARED, POI_NOT_SHARED, POI_SELF_SHARE, POI_TRANSFER_SELF, ORG_PERMISSION_REQUIRED, INTERNAL_ERROR]
        message: { type: string }
        status: { type: integer }

    # ── Enums ──
    HolderType:
      type: string
      enum: [player, npc, corporation, system]
...
    InstanceStatus:
      type: string
      enum: [stored, in_world]
...
    PoiOwnerType:
      type: string
      enum: [player, npc, corporation, political, system]
    PoiGranteeType:
      type: string
      enum: [player, npc, corporation, political]
    PoiVisibility:
      type: string
      enum: [private, public]

    # ── Model ──
...
    Stack:
      type: object
      required: [holderType, holderId, goodType, quantity, updatedAt]
      properties:
        holderType: { $ref: '#/components/schemas/HolderType' }
        holderId: { type: string, format: uuid }
        goodType: { type: string }
        quantity: { type: integer, description: Total owned quantity (integer units) }
        updatedAt: { type: string, format: date-time }
    StackView:
      allOf:
        - $ref: '#/components/schemas/Stack'
        - type: object
          required: [held, available]
          properties:
            held: { type: integer, description: Quantity reserved by active holds }
            available: { type: integer, description: "`quantity - held`, the tradable/consumable part" }
    Instance:
      type: object
      required: [id, goodType, holderType, holderId, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, description: Stable instance UUID }
        goodType: { type: string }
        holderType: { $ref: '#/components/schemas/HolderType' }
        holderId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/InstanceStatus' }
        metadata: { type: [object, 'null'], additionalProperties: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    HolderInventory:
      type: object
      required: [holderType, holderId, stacks, instances]
      properties:
        holderType: { $ref: '#/components/schemas/HolderType' }
        holderId: { type: string, format: uuid }
        stacks:
          type: array
          items: { $ref: '#/components/schemas/StackView' }
        instances:
          type: array
          items: { $ref: '#/components/schemas/Instance' }
...
    Poi:
      type: object
      description: >
        A GPS point, optionally widened into a zone by `radiusM`. Owned by exactly one holder
        and readable by its owner, its grantees and (when `public`) every player.
      required: [id, ownerType, ownerId, name, x, y, z, visibility, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        ownerType: { $ref: '#/components/schemas/PoiOwnerType' }
        ownerId: { type: string, format: uuid }
        name: { type: string, maxLength: 128 }
        description: { type: [string, 'null'], maxLength: 2000 }
        system: { type: [string, 'null'], maxLength: 64, description: Star system (null = unscoped) }
        scene:
          type: [string, 'null']
          maxLength: 128
          description: Hierarchical scene path matched by equality or segment prefix (`tarsis_1` matches `tarsis_1/new-paris`, never `tarsis_10`)
        parentId:
          type: [string, 'null']
          format: uuid
          description: Opaque parent object id in Persistence (no cross-service foreign key)
        x: { type: number, description: Game units (metres assumed) }
        y: { type: number }
        z: { type: number }
        radiusM:
          type: [number, 'null']
          minimum: 0
          maximum: 1000000
          description: '`null` = plain point (exact position); `>= 0` = zone radius in metres'
        visibility: { $ref: '#/components/schemas/PoiVisibility' }
        metadata: { type: [object, 'null'], additionalProperties: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PoiShare:
      type: object
      description: Read-only grant of a POI (the grantee never gets write access).
      required: [poiId, granteeType, granteeId, createdAt]
      properties:
        poiId: { type: string, format: uuid }
        granteeType: { $ref: '#/components/schemas/PoiGranteeType' }
        granteeId: { type: string, format: uuid }
        sharedBy: { type: [string, 'null'], format: uuid, description: Player who created the grant (audit) }
        createdAt: { type: string, format: date-time }
    PoiView:
      allOf:
        - $ref: '#/components/schemas/Poi'
        - type: object
          required: [shares]
          properties:
            shares:
              type: array
              items: { $ref: '#/components/schemas/PoiShare' }
...